+++ src/main/kotlin/com/github/jeancsanchez/leaofaminto/domain/model/Venda.kt
@@ -18,7 +18,7 @@ class Venda(
     corretora: Corretora,
     data: LocalDate = LocalDate.now(),
     preco: Double,
-    quantidade: Int
+    quantidade: Double
 ) : Operacao(ativo, corretora, data, preco, quantidade) {
 
     @Enumerated(EnumType.STRING)
